Refactor TOSAOperatorSupport to allow for maping against TosaSpecification
