+++ backends/arm/operator_support/__init__.py
@@ -1,0 +1,8 @@
+# Copyright 2024 Arm Limited and/or its affiliates.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+# pyre-unsafe
+
+from . import mean_dim_support, tosa_supported_operators, var_correction_support  # noqa
